Refactor Characters
