+++ src/Library/Characters/Knight.cs
@@ -1,90 +1,8 @@
-using System.Collections.Generic;
-
 namespace RoleplayGame
 {
-    public class Knight : ICharacter
+    public class Knight : Character
     {
-        private int health = 100;
 
-        public Knight(string name)
-        {
-            this.Name = name;
-        }
-
-        public string Name { get; set; }
-
-        private IList<IDefensiveItems> defensiveEquipment = new List<IDefensiveItems>();
-
-        private IList<IAttackItems> offensiveEquipment = new List<IAttackItems>();
-
-        public void EquipDefensiveItem(IDefensiveItems defensiveItem)
-        {
-            this.defensiveEquipment.Add(defensiveItem);
-        }
-
-        public void UnEquipDefensiveItem(IDefensiveItems defensiveItem)
-        {
-            this.defensiveEquipment.Remove(defensiveItem);
-        }
-
-        public void EquipAttackItem(IAttackItems attackItem)
-        {
-            this.offensiveEquipment.Add(attackItem);
-        }
-
-        public void UnEquipAttackItem(IAttackItems attackItem)
-        {
-            this.offensiveEquipment.Remove(attackItem);
-        }
-
-        public int GetTotalAttackValue()
-        {   
-            int result = 0;
-            foreach (IAttackItems attackItem in offensiveEquipment)
-            {
-                result += attackItem.AttackValue();
-            }
-            return result;
-        }
-
-        public int GetTotalDefenseValue()
-        {
-            int result = 0;
-            foreach (IDefensiveItems defensiveItem in defensiveEquipment)
-            {
-                result += defensiveItem.DefenseValue();
-            }
-            return result;
-        }
-
-        public int Health
-        {
-            get
-            {
-                return this.health;
-            }
-            private set
-            {
-                this.health = value < 0 ? 0 : value;
-            }
-        }
-
-        public int ReceiveAttack(int damage)
-        {
-            if (damage - this.GetTotalDefenseValue() > 0)
-            {
-                this.Health = this.Health - (damage - this.GetTotalDefenseValue());
-                return this.Health;
-            }
-            else
-            {
-                return this.Health;
-            }
-        }
-
-        public int Cure()
-        {
-            return this.Health = 100;
-        }
+        public Knight(string name) : base(name) { }
     }
 }
