+++ src/Library/Characters/Wizard.cs
@@ -2,54 +2,11 @@ using System.Collections.Generic;
 
 namespace RoleplayGame
 {
-    public class Wizard : ICharacter
+    public class Wizard : Character
     {
-        private int health = 100;
-
-        public Wizard(string name)
-        {
-            this.Name = name;
-        }
-
-        public string Name { get; set; }
-
-        private IList<IDefensiveItems> defensiveEquipment = new List<IDefensiveItems>();
-
-        private IList<IAttackItems> offensiveEquipment = new List<IAttackItems>();
-
-        private IList<IMixedItems> mixedEquipment = new List<IMixedItems>();
+        public Wizard(string name) : base(name) { }
 
         private IList<SpellsBook> spellsBookList = new List<SpellsBook>();
-
-        public void EquipDefensiveItem(IDefensiveItems defensiveItem)
-        {
-            this.defensiveEquipment.Add(defensiveItem);
-        }
-
-        public void UnEquipDefensiveItem(IDefensiveItems defensiveItem)
-        {
-            this.defensiveEquipment.Remove(defensiveItem);
-        }
-
-        public void EquipAttackItem(IAttackItems attackItem)
-        {
-            this.offensiveEquipment.Add(attackItem);
-        }
-
-        public void UnEquipAttackItem(IAttackItems attackItem)
-        {
-            this.offensiveEquipment.Remove(attackItem);
-        }
-
-        public void EquipMixedItem(IMixedItems mixedItem)
-        {
-            this.mixedEquipment.Add(mixedItem);
-        }
-
-        public void UnEquipMixedItem(IMixedItems mixedItem)
-        {
-            this.mixedEquipment.Remove(mixedItem);
-        }
 
         public void EquipSpellBook(SpellsBook spellsBook)
         {
@@ -64,7 +21,7 @@ namespace RoleplayGame
             this.spellsBookList.Remove(spellsBook);
         }
 
-        public int GetTotalAttackValue()
+        public override int GetTotalAttackValue()
         {
             int result = 0;
             foreach (IAttackItems attackItems in offensiveEquipment)
@@ -85,7 +42,7 @@ namespace RoleplayGame
             return result;
         }
 
-        public int GetTotalDefenseValue()
+        public override int GetTotalDefenseValue()
         {
             int result = 0;
             foreach (IDefensiveItems defensiveItem in defensiveEquipment)
@@ -104,36 +61,6 @@ namespace RoleplayGame
                 }
             }
             return result;
-        }
-
-        public int Health
-        {
-            get
-            {
-                return this.health;
-            }
-            private set
-            {
-                this.health = value < 0 ? 0 : value;
-            }
-        }
-
-        public int ReceiveAttack(int damage)
-        {
-            if (damage - this.GetTotalDefenseValue() > 0)
-            {
-                this.Health = this.Health - (damage - this.GetTotalDefenseValue());
-                return this.Health;
-            }
-            else
-            {
-                return this.Health;
-            }
-        }
-
-        public int Cure()
-        {
-            return this.Health = 100;
         }
     }
 }
